删除语句增加空校验，防止误删除tomca下所有的目录

+++ docker/tools/remove.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 app=$1
 
+if [ ! $app ]
+then 
+  echo "App is null"
+  exit 0
+fi
+
 sudo docker stop $app
 sudo docker rm $app
